refactor: extract Renderer + FrameState into compilertop package
Splits the render-side of CompilerTop into its own file. The new
`class Renderer` owns the rolling thread-occupancy history (the only
mutable state on the render path); everything else flows in via a
new top-level `FrameState` case class assembled by CompilerTop once
per tick. Renderer is now a pure function of state and can be
exercised without a live `Flix`.

CompilerTop shrinks from 674 to ~320 lines and is now scoped to
lifecycle + snapshotting (threads, JLine, raw mode, atomics,
frozen* values, start/stop/inputLoop/loop). The dashboard /
sparkline constants (TotalPhases, MaxPhaseLen, BarWidth,
ProgressStartCol, ThreadsStartCol, SparkWidth, SparkChars) move
to Renderer's companion.

Subtle improvement: `buildFrameState()` reads `completed.get()` once
and uses it consistently for elapsed / threads / heap. Previously
`renderStats` re-read it independently, which could in principle
flip between reads within a single frame.

Pure code move + a tiny consistency fix; no observable behavior
change.
